requirements.txt
netifaces==0.11.0
paho-mqtt==2.1.0
pycryptodome==3.20.0
